Split ixmp/backend/jdbc.py to 3 files
- jdbc/
  - __init__.py contains the JDBCBackend class only.
  - jvm.py contains code for interacting with the JVM, ixmp_source,
    JPype, etc.
  - options.py contains the Options class.
- Add JClassProxy and use more explicit paths to reference Java classes.
- Expand and improve docstrings.
- Adjust imports.
- Reorganize tests.
